Drop memory allocation and non-const methods
This commit drops all the parts of the code related to memory allocation
and those performing changes on the DOM. In fact, we intend to preserve
the DOM, but in read-only mode.

Note that in certain points TODOs have been left, where the
implementations of methods we want to keep will require some attention
to decide how to handle them.